fix: refactor some of the functions
